requirements.txt
python-telegram-bot==13.1
transformers
diffusers[torch]==0.12.1
openai